Create Rock, Paper, Scissors
A set of code to create a game of rock, paper, scissors using Python from Codecademy. I'm trying to make it to where I don't have to keep typing python RPS.py in a terminal to bring up the "Enter Rock, Paper, or Scissors: " prompt. I want the prompt to be automatically printed after the user wins, loses, or ties with the computer.